Extract classes into their own files and create Logger.get_logger() to get a logger
